split out router into moules
